Refactor chip control: use blocking delays and consolidate modules
Chip control refactoring:
- Move NetControl and UiControl impl blocks from chip_control.rs to gpio.rs
  (keep implementations with struct definitions)
- Remove chip_control.rs module entirely
- Make power_cycle a private method on NetControl (only caller)

Replace async delays with blocking delays:
- Change from embassy_time::Timer (async) to Delay (blocking)
- Remove async/await from all chip control methods
- Update all callers to remove .await

Benefits:
- Simpler code - no async overhead for simple delays
- Better module organization - impls with structs
- Fewer files to navigate

All builds succeed.
